it seems all i need was a simple queue
First of all instead of maintaining the active page instances to a
fixed number i start 10 page instances at once and Promise.all them.
after i do that i take 10 elements from the queue and do this again
until the queue is empty. You can see the while loop in the crawl
methoad.

Second for some urls i was getting
[Error]
the execution context was destroyed, probably because of a navigation.
The urls at which i was getting this error was consistent.
Turns out these urls had a 301 response code(redirect).
And i was trying to execute page.$$eval function on the initial
response page. For now i just catch the error and move on.
But i need to modify the visit function to be more error resillient.